% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envvar.R
\name{envvar}
\alias{envvar}
\title{Look up an environment variable}
\usage{
envvar(name, mode = "character")
}
\arguments{
\item{name}{The name of the environment variable.  Typically this
is an upper-case name, with words separated by underscores}

\item{mode}{Mode to set the variable to.  This can be one of
"character" (the default) or "integer"}
}
\description{
Look up an environment variable.  This is a wrapper around
\code{\link{Sys.getenv}} but designed to ease non-interactive use
of vault.
}
