% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_secrets.R
\name{resolve_secrets}
\alias{resolve_secrets}
\title{Use vault to resolve secrets.}
\usage{
resolve_secrets(x, ..., login = TRUE)
}
\arguments{
\item{x}{List of properties values of which may be vault secrets of the form
VAULT:<path to secret>:<key to get>. Any values which don't match pattern
of a secret are ignored and returned as is.}

\item{...}{Args to be passed to \code{\link{vault_client}} call.}

\item{login}{Login method to be passed to call to \code{\link{vault_client}}.}
}
\value{
List of properties with any vault secrets resolved.
}
\description{
Sets up a vault client via call to \code{\link{vault_client}} and resolves
secrets specified by the form VAULT:<path to secret>:<key to get>.
}
\examples{
\dontrun{
srv <- vaultr::vault_test_server()
cl <- srv$client()
cl$write("/secret/users/alice", list(password = "ALICE"))
x <- list(name = "alice",
  password = "VAULT:/secret/users/alice:password")
withr::with_envvar(c(VAULTR_AUTH_METHOD = "token", VAULT_TOKEN = srv$token),
  x <- resolve_secrets(x, addr = srv$addr)
)
}


}
