% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_auth}
\alias{vault_auth}
\title{Authenticate vault}
\usage{
vault_auth(addr = NULL, renew = FALSE, quiet = FALSE)
}
\arguments{
\item{addr}{Optional vault address, to override the
\code{VAULT_ADDR} environment variable.}

\item{renew}{Re-authenticate, even if a token is already found}

\item{quiet}{Don't print information on success}
}
\description{
Authenticate to vault.  Currently the only authentication method
supported is github.
}
\details{
You may need to set the following environment variables

\itemize{

\item{\code{VAULT_ADDR}: The address that your vault server is running on}

\item{\code{VAULT_AUTH_GITHUB_TOKEN}: The github token that you
have set up to work with vault}

}
}
