% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_cache.R
\name{vault_clear_token_cache}
\alias{vault_clear_token_cache}
\title{Clear vaultr token cache}
\usage{
vault_clear_token_cache(session = TRUE, persistent = TRUE,
  cache_dir = NULL, quiet = FALSE)
}
\arguments{
\item{session}{Logical, indicating if we should clear the session
keys}

\item{persistent}{Logical, indicating if we should clear the
persistent keys}

\item{cache_dir}{Location of persistent keys (otherwise uses the
\code{VAULTR_CACHE_DIR} environment variable)}

\item{quiet}{Logical, indicating if messages should be ommited}
}
\description{
Clear all cached token
}
\details{
Vaultr stores tokens in two places - one session specific and one
persistent.  This function clears these out which might be useful
in debugging (otherwise pass in the \code{renew} argument to
\code{auth}).  It also comes in useful in testing packages that
depend on vaultr.
}
