% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{vault_client}
\alias{vault_client}
\title{Make a vault client}
\usage{
vault_client(addr = NULL, auth = NULL, ..., cert = NULL, verify = NULL)
}
\arguments{
\item{addr}{Vault address, e.g. \code{https://localhost:8200}.  If
not given, defaults to the environment variable
\code{VAULT_ADDR}.}

\item{auth}{An authentication method (e.g., "github")}

\item{...}{Additional arguments passed through to the auth method}

\item{cert}{Client certificate}

\item{verify}{Server certificate}
}
\description{
Make a vault client
}
