% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{vault_client_generic}
\alias{vault_client_generic}
\title{Make a generic vault client}
\usage{
vault_client_generic(...)
}
\arguments{
\item{...}{Passed to \code{vault_client}}
}
\description{
Make a vault client for the "generic" interface.  Provides a much
simpler interface than \code{vault_client}, with none of the
administration functions.
}
