% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_list}
\alias{vault_list}
\title{List keys}
\usage{
vault_list(path, recursive = FALSE)
}
\arguments{
\item{path}{Vault path to list keys in (e.g., "secret")}

\item{recursive}{List keys recursively?}
}
\description{
List keys in the vault
}
