% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_read}
\alias{vault_read}
\title{Read values from a key}
\usage{
vault_read(path, field = NULL)
}
\arguments{
\item{path}{Vault path to read}

\item{field}{Field to read (default returns all fields as a list)}
}
\description{
Read the values from a key
}
