% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_manager.R
\name{vault_test_server_start}
\alias{vault_test_server_start}
\alias{vault_test_server_stop}
\alias{vault_test_server}
\alias{vault_test_client}
\alias{vault_test_server_install}
\title{Control a test vault server}
\usage{
vault_test_server_start()

vault_test_server_stop()

vault_test_server()

vault_test_client(...)

vault_test_server_install(path = ".vault", quiet = FALSE,
  version = "0.7.3")
}
\arguments{
\item{...}{Argument passed through to create the new client}

\item{path}{Path to install the server to; must be an existing
directory.}

\item{quiet}{Suppress progress bars on install}

\item{version}{Version of vault to install}
}
\description{
Control a server for use with testing.  This is designed to be
used only by other packages that wish to run tests against a vault
server.
}
\details{
The function \code{vault_test_server_install} will install a test
server, but \emph{only} if the user opts in by setting the
environment variable \code{VAULTR_TEST_SERVER_INSTALL} to
\code{"true"}.  This will download a ~50MB binary from
\url{https://vaultproject.io} so use with care.  It is intended
\emph{only} for use in automated testing environments.
}
