% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_client.R
\name{vault_client}
\alias{vault_client}
\title{Make a vault client}
\usage{
vault_client(login = FALSE, ..., addr = NULL, tls_config = NULL)
}
\arguments{
\item{login}{Login method.  Specify a string to be passed along as
the \code{method} argument to \code{$login}.  The default
\code{FALSE} means not to login.  \code{TRUE} means to login
using a default method specified by the environment variable
\code{VAULTR_AUTH_METHOD} - if that variable is not set, an
error is thrown.  The value of \code{NULL} is the same as
\code{TRUE} but does not throw an error if
\code{VAULTR_AUTH_METHOD} is not set.  Supported methods are
\code{token}, \code{github} and \code{userpass}.}

\item{...}{Additional arguments passed along to the authentication
method indicated by \code{login}, if used.}

\item{addr}{The value address \emph{including protocol and port},
e.g., \code{https://vault.example.com:8200}.  If not given, the
default is the environment variable \code{VAULT_ADDR}, which is
the same as used by vault's command line client.}

\item{tls_config}{TLS (https) configuration.  For most uses this
can be left blank.  However, if your vault server uses a
self-signed certificate you will need to provide this.  Defaults
to the environment variable \code{VAULT_CAPATH}, which is the
same as vault's command line client.}
}
\description{
Make a vault client.  This must be done before accessing the
vault.  The deafults for arguments are controlled by environment
variables (see Details) and values provided as arguments override
these defaults.
}
\author{
Rich FitzJohn
}
