% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{vault_client}
\alias{vault_client}
\title{Make a vault client}
\usage{
vault_client(auth_method = NULL, ..., addr = NULL, verify = NULL)
}
\arguments{
\item{auth_method}{An authentication method (e.g., "token",
"github").  If \code{NULL}, we try the value of
\code{VAULTR_AUTH_METHOD}, and if that is not set then
authentication is not done.}

\item{...}{Additional arguments passed through to the auth method}

\item{addr}{Vault address, e.g. \code{https://localhost:8200}.  If
not given, defaults to the environment variable
\code{VAULT_ADDR}}

\item{verify}{Server certificate (or \code{VAULT_CAPATH})}
}
\description{
Make a vault client.  The defaults for arguments are controlled by
environment variables.
}
