% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault_client.R
\name{vault_client}
\alias{vault_client}
\alias{R6_vault_client}
\title{Make a vault client}
\usage{
vault_client(login = FALSE, ..., addr = NULL, tls_config = NULL)
}
\arguments{
\item{login}{Login method.  Specify a string to be passed along as
the \code{method} argument to \code{$login}.  The default
\code{FALSE} means not to login.  \code{TRUE} means to login
using a default method specified by the environment variable
\code{VAULTR_AUTH_METHOD} - if that variable is not set, an
error is thrown.  The value of \code{NULL} is the same as
\code{TRUE} but does not throw an error if
\code{VAULTR_AUTH_METHOD} is not set.  Supported methods are
\code{token}, \code{github} and \code{userpass}.}

\item{...}{Additional arguments passed along to the authentication
method indicated by \code{login}, if used.}

\item{addr}{The value address \emph{including protocol and port},
e.g., \code{https://vault.example.com:8200}.  If not given, the
default is the environment variable \code{VAULT_ADDR}, which is
the same as used by vault's command line client.}

\item{tls_config}{TLS (https) configuration.  For most uses this
can be left blank.  However, if your vault server uses a
self-signed certificate you will need to provide this.  Defaults
to the environment variable \code{VAULT_CAPATH}, which is the
same as vault's command line client.}
}
\description{
Make a vault client.  This must be done before accessing the
vault.  The default values for arguments are controlled by
environment variables (see Details) and values provided as
arguments override these defaults.
}
\section{Environment variables}{


The creation of a client is affected by a number of environment
variables, following the main vault command line client.

\describe{

\item{\code{VAULT_ADDR}}{The url of the vault server.  Must
include a protocol (most likely \code{https://} but in testing
\code{http://} might be used)}

\item{\code{VAULT_CAPATH}}{The path to CA certificates}

\item{\code{VAULT_TOKEN}}{A vault token to use in authentication.
Only used for token-based authentication}

\item{\code{VAULT_AUTH_GITHUB_TOKEN}}{As for the command line
client, a github token for authentication using the github
authentication backend}

\item{\code{VAULTR_AUTH_METHOD}}{The method to use for
authentication}

}
}

\examples{


# We work with a test vault server here (see ?vault_test_server) for
# details.  To use it, you must have a vault binary installed on your
# system.  These examples will not affect any real running vault
# instance that you can connect to.
server <- vaultr::vault_test_server(if_disabled = message)

if (!is.null(server)) {
  # Create a vault_client object by providing the address of the vault
  # server.
  client <- vaultr::vault_client(addr = server$addr)

  # The client has many methods, grouped into a structure:
  client

  # For example, token related commands:
  client$token

  # The client is not authenticated by default:
  try(client$list("/secret"))

  # A few methods are unauthenticated and can still be run
  client$status()

  # Login to the vault, using the token that we know from the server -
  # ordinarily you would use a login approach suitable for your needs
  # (see the vault documentation).
  token <- server$token
  client$login(method = "token", token = token)

  # The vault contains no secrets at present
  client$list("/secret")

  # Secrets can contain any (reasonable) number of key-value pairs,
  # passed in as a list
  client$write("/secret/users/alice", list(password = "s3cret!"))

  # The whole list can be read out
  client$read("/secret/users/alice")
  # ...or just a field
  client$read("/secret/users/alice", "password")

  # Reading non-existant values returns NULL, not an error
  client$read("/secret/users/bob")

  client$delete("/secret/users/alice")
}
}
\author{
Rich FitzJohn
}
\section{Super class}{
\code{\link[vaultr:vault_client_object]{vaultr::vault_client_object}} -> \code{vault_client}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{R6_vault_client$new()}}
\item \href{#method-api}{\code{R6_vault_client$api()}}
\item \href{#method-read}{\code{R6_vault_client$read()}}
\item \href{#method-write}{\code{R6_vault_client$write()}}
\item \href{#method-delete}{\code{R6_vault_client$delete()}}
\item \href{#method-list}{\code{R6_vault_client$list()}}
\item \href{#method-login}{\code{R6_vault_client$login()}}
\item \href{#method-status}{\code{R6_vault_client$status()}}
\item \href{#method-unwrap}{\code{R6_vault_client$unwrap()}}
\item \href{#method-wrap_lookup}{\code{R6_vault_client$wrap_lookup()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="vaultr" data-topic="vault_client_object" data-id="format">}\href{../../vaultr/html/vault_client_object.html#method-format}{\code{vaultr::vault_client_object$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="vaultr" data-topic="vault_client_object" data-id="help">}\href{../../vaultr/html/vault_client_object.html#method-help}{\code{vaultr::vault_client_object$help()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$new(addr, tls_config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-api"></a>}}
\if{latex}{\out{\hypertarget{method-api}{}}}
\subsection{Method \code{api()}}{
Returns an api client object that can be used to
directly interact with the vault server.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$api()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-read"></a>}}
\if{latex}{\out{\hypertarget{method-read}{}}}
\subsection{Method \code{read()}}{
Read a value from the vault.  This can be used to
read any value that you have permission to read, and can also
be used as an interface to a version 1 key-value store (see
\link{vault_client_kv1}.  Similar to the vault CLI command
\verb{vault read}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$read(path, field = NULL, metadata = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path for the secret to read, such as
\verb{/secret/mysecret}}

\item{\code{field}}{Optional field to read from the secret.  Each
secret is stored as a key/value set (represented in R as a
named list) and this is equivalent to using \verb{[[field]]} on
the return value.  The default, \code{NULL}, returns the full set
of values.}

\item{\code{metadata}}{Logical, indicating if we should return
metadata for this secret (lease information etc) as an
attribute along with the values itself.  Ignored if \code{field}
is specified.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write"></a>}}
\if{latex}{\out{\hypertarget{method-write}{}}}
\subsection{Method \code{write()}}{
Write data into the vault.  This can be used to
write any value that you have permission to write, and can
also be used as an interface to a version 1 key-value store
(see \link{vault_client_kv1}.  Similar to the vault CLI
command \verb{vault write}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$write(path, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path for the secret to write, such as
\verb{/secret/mysecret}}

\item{\code{data}}{A named list of values to write into the vault at
this path.  This \emph{replaces} any existing values.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete a value from the vault
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$delete(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to delete}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-list"></a>}}
\if{latex}{\out{\hypertarget{method-list}{}}}
\subsection{Method \code{list()}}{
List data in the vault at a given path.  This can
be used to list keys, etc (e.g., at \verb{/secret}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$list(path, full_names = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The path to list}

\item{\code{full_names}}{Logical, indicating if full paths (relative
to the vault root) should be returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character vector (of zero length if no keys are
found).  Paths that are "directories" (i.e., that contain
keys and could themselves be listed) will be returned with a
trailing forward slash, e.g. \verb{path/}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-login"></a>}}
\if{latex}{\out{\hypertarget{method-login}{}}}
\subsection{Method \code{login()}}{
Login to the vault.  This method is more
complicated than most.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$login(
  ...,
  method = "token",
  mount = NULL,
  renew = FALSE,
  quiet = FALSE,
  token_only = FALSE,
  use_cache = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Additional named parameters passed through to the
underlying method}

\item{\code{method}}{Authentication method to use, as a string.
Supported values include \code{token} (the default), \code{github},
\code{approle} and \code{userpass}.}

\item{\code{mount}}{The mount path for the authentication backend, \emph{if
it has been mounted in a nonstandard location}.  If not
given, then it is assumed that the backend was mounted at a
path corresponding to the method name.}

\item{\code{renew}}{Login, even if we appear to hold a valid token.
If \code{FALSE} and we have a token then \code{login} does nothing.}

\item{\code{quiet}}{Suppress some informational messages}

\item{\code{token_only}}{Logical, indicating that we do not want to
actually log in, but instead just generate a token and return
that.  IF given then \code{renew} is ignored and we always
generate a new token.}

\item{\code{use_cache}}{Logical, indicating if we should look in the
session cache for a token for this client.  If this is \code{TRUE}
then when we log in we save a copy of the token for this
session and any subsequent calls to \code{login} at this vault
address that use \code{use_cache = TRUE} will be able to use this
token.  Using cached tokens will make using some
authentication backends that require authentication with
external resources (e.g., \code{github}) much faster.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-status"></a>}}
\if{latex}{\out{\hypertarget{method-status}{}}}
\subsection{Method \code{status()}}{
Return the status of the vault server, including
whether it is sealed or not, and the vault server version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-unwrap"></a>}}
\if{latex}{\out{\hypertarget{method-unwrap}{}}}
\subsection{Method \code{unwrap()}}{
Returns the original response inside the given
wrapping token. The vault endpoints used by this method
perform validation checks on the token, returns the original
value on the wire rather than a JSON string representation of
it, and ensures that the response is properly audit-logged.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$unwrap(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{Specifies the wrapping token ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-wrap_lookup"></a>}}
\if{latex}{\out{\hypertarget{method-wrap_lookup}{}}}
\subsection{Method \code{wrap_lookup()}}{
Look up properties of a wrapping token.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{R6_vault_client$wrap_lookup(token)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{Specifies the wrapping token ID to lookup}
}
\if{html}{\out{</div>}}
}
}
}
