% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_manager.R
\name{vault_test_server}
\alias{vault_test_server}
\alias{vault_test_server_install}
\title{Control a test vault server}
\usage{
vault_test_server(https = FALSE, init = TRUE,
  if_disabled = testthat::skip)

vault_test_server_install(quiet = FALSE, version = "0.10.3")
}
\arguments{
\item{https}{Logical scalar, indicating if a https-using server
should be created, rather than the default vault dev-mode
server.  This is still \emph{entirely} insecure, and uses self
signed certificates that are bundled with the package.}

\item{init}{Logical scalar, indicating if the https-using server
should be initialised.}

\item{if_disabled}{Callback function to run if the vault server is
not enabled.  The default, designed to be used within tests, is
\code{testthat::skip}.  Alternatively, inspect the
\code{$enabled} property of the returned object.}

\item{quiet}{Suppress progress bars on install}

\item{version}{Version of vault to install}
}
\description{
Control a server for use with testing.  This is designed to be
used only by other packages that wish to run tests against a vault
server.  You will need to set \code{VAULTR_VAULT_BIN_PATH} to
point at the directory containing the vault binary.
}
\details{
The function \code{vault_test_server_install} will install a test
server, but \emph{only} if the user opts in by setting the
environment variable \code{VAULTR_TEST_SERVER_INSTALL} to
\code{"true"}, and by setting \code{VAULTR_VAULT_BIN_PATH} to the
directory where the binary should be downloaded to.  This will
download a ~100MB binary from \url{https://vaultproject.io} so use
with care.  It is intended \emph{only} for use in automated
testing environments.
}
