% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_manager2.R
\name{vault_test_server}
\alias{vault_test_server}
\alias{vault_test_server_install}
\title{Control a test vault server}
\usage{
vault_test_server(https = FALSE)

vault_test_server_install(quiet = FALSE, version = "0.10.3")
}
\arguments{
\item{quiet}{Suppress progress bars on install}

\item{version}{Version of vault to install}
}
\description{
Control a server for use with testing.  This is designed to be
used only by other packages that wish to run tests against a vault
server.  You will need to set \code{VAULT_BIN_PATH} to point at
the directory containing the vault binary.
}
\details{
The function \code{vault_test_server_install} will install a test
server, but \emph{only} if the user opts in by setting the
environment variable \code{VAULTR_TEST_SERVER_INSTALL} to
\code{"true"}, and by setting \code{VAULT_BIN_PATH} to the
directory where the binary should be downloaded to.  This will
download a ~100MB binary from \url{https://vaultproject.io} so use
with care.  It is intended \emph{only} for use in automated
testing environments.
}
